More memory optimizations

+++ Gui/Elements/Pager.php
@@ -12,17 +12,17 @@ use ManiaLivePlugins\eXpansion\Gui\Config;
 class Pager extends \ManiaLivePlugins\eXpansion\Gui\Control implements \ManiaLivePlugins\eXpansion\Gui\Structures\ScriptedContainer
 {
 
-	private $pager;
+	protected $pager;
 
-	private $items = array();
+	protected $items = array();
 
-	private $scroll;
+	protected $scroll;
 
-	private $scrollBg, $scrollUp, $scrollDown, $barFrame;
+	protected $scrollBg, $scrollUp, $scrollDown, $barFrame;
 
-	private $itemSizeY = 6;
+	protected $itemSizeY = 6;
 
-	private $myScript;
+	protected $myScript;
 
 	public function __construct()
 	{
@@ -120,11 +120,10 @@ class Pager extends \ManiaLivePlugins\eXpansion\Gui\Control implements \ManiaLiv
 
 	public function clearItems()
 	{
-		foreach ($this->items as $item) {
-			$this->pager->removeComponent($item);
-			$item->destroy();
+		if (property_exists($this, "pager")) {
+			$this->pager->destroyComponents();
+			$this->items = array();
 		}
-		$this->items = array();
 	}
 
 	public function removeItem(\ManiaLib\Gui\Component $item)
@@ -137,8 +136,10 @@ class Pager extends \ManiaLivePlugins\eXpansion\Gui\Control implements \ManiaLiv
 
 	public function destroy()
 	{
-		$this->clearItems();
-		$this->pager->destroy();
+		if (property_exists($this, "pager")) {
+			$this->pager->destroyComponents();
+			$this->pager->destroy();
+		}
 		parent::destroy();
 	}
